move transformation test to test_transformations.py
